requirements.txt
alabaster==0.7.13
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.14.1
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.80
black==23.1.0
bleach==4.1.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
configparser==5.3.0
coverage==7.1.0
cryptography==38.0.4
csvsort==1.6.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docutils==0.19
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
flake8==6.0.0
flit_core==3.6.0
future==0.18.3
icecream==2.1.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.2.0
iniconfig==2.0.0
ipykernel==6.19.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.12.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.1
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.5
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.9.0
loguru==0.6.0
lxml==4.9.1
mando==0.6.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
mypy==0.991
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
progressbar2==4.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.16.0
pyOpenSSL==22.0.0
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.4.0
QtPy==2.2.0
radon==5.1.0
requests==2.28.2
rna-tools==3.12.1
scipy==1.10.0
Send2Trash==1.8.0
setuptools==65.6.3
sip==6.6.2
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-autoprogram==0.1.7
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.14
versioneer==0.28
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wget==3.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.1
zipp==3.12.1
